Added OneStepAheadStrategy and SimulatedMoveResult.
We now first check if we cant reach our last treasure but an enemy could
reach his last treasure!
Then we filter all shift operations such that the enemy cant reach his
last treasure (it can happen, that indepent from our shift the enemy can
reach his last treasure).
After that we test if we can find our move and if not, we use the
OneStepAheadStrategy.
I test the OneStepAheadStrategy against the ClosestToTreasureStrategy
and they seam to be equally good (or bad).